# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for KNS Indexer",
        "title": "KNS Indexer API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/primary-username/{owner}": {
            "get": {
                "description": "Returns primary username by owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owner"
                ],
                "summary": "Resolve primary username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetPrimaryUsernameSuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    }
                }
            }
        },
        "/usernames": {
            "get": {
                "description": "Returns paginated list of all registered usernames with sorting by timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "username"
                ],
                "summary": "Get list of usernames",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 100,
                        "description": "Number of records per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination (starts from 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order by timestamp: asc or desc",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved usernames",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetUsernamesSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    }
                }
            }
        },
        "/usernames/owner/{owner}": {
            "get": {
                "description": "Returns paginated list of registered usernames by owner with sorting by timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "owner"
                ],
                "summary": "Get list of owner usernames",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 100,
                        "description": "Number of records per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination (starts from 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order by timestamp: asc or desc",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved usernames",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetOwnerUsernamesSuccessResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    }
                }
            }
        },
        "/usernames/{username}": {
            "get": {
                "description": "Returns username record by username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "username"
                ],
                "summary": "Resolve username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetUsernameSuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.GetOwnerUsernamesSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.GetOwnerUsernamesSuccessResponseData"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handlers.GetOwnerUsernamesSuccessResponseData": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "usernames": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Username"
                    }
                }
            }
        },
        "handlers.GetPrimaryUsernameSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Username"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handlers.GetUsernameSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Username"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handlers.GetUsernamesSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.GetUsernamesSuccessResponseData"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handlers.GetUsernamesSuccessResponseData": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "usernames": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Username"
                    }
                }
            }
        },
        "models.FailureResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "description"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "models.Username": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "keeta_aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                },
                "cid": {
                    "type": "string",
                    "example": "Qmaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                },
                "isPrimary": {
                    "type": "boolean",
                    "example": false
                },
                "owner": {
                    "type": "string",
                    "example": "keeta_bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2025-11-25T11:22:33.123Z"
                },
                "username": {
                    "type": "string",
                    "example": "username"
                }
            }
        }
    }
}